+++ sample/proto_v2/csv/KindConst.h
@@ -18,9 +18,9 @@ struct FKindConst : public FTableRowBase
     UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "XResConfig")
     FName Name;
 
-    /** Field Type: INT, Name: Value. This field is generated for UE Editor compatible. **/
+    /** Field Type: STRING, Name: Value. This field is generated for UE Editor compatible. **/
     UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "XResConfig")
-    int32 Value;
+    FString Value;
 
 };